update the README and rename the file name
