+++ README.md
@@ -1,1 +1,7 @@
 # Futures
+Analyze the trend of Taiwan futures to facilitate investors to make reference to historical data for layout
+
+## A brief introduction to job content
+- Use Selenium and Chromedriver to crawl page data
+- Store crawled data in Mongo DB
+- Use Matplotlib for data visualization to complete the presentation of 5MA, 10MA, and K-line charts
